+++ src/app/map/app-map-layer/app-map-layer/shared-mobility/shared-mobility.ts
@@ -1,5 +1,4 @@
-import OJP_Legacy from '../../../../config/ojp-legacy';
-
+import { OJP_VERSION } from '../../../../config/constants';
 import { AnyPlace } from '../../../../shared/models/place/place-builder';
 import { Poi, RestrictionPoiOSMTag } from '../../../../shared/models/place/poi';
 
@@ -63,112 +62,114 @@ export class SharedMobility {
 
     // see https://github.com/SFOE/sharedmobility/blob/main/providers.csv
     const providerData: [SharedMobilityProvider | null, VehicleType] = (() => {
+      if (OJP_VERSION === '1.0') {
+        const placeName = (poi.placeName ?? 'n/a').toLowerCase();
+        if (placeName.startsWith('bolt')) {
+          return ['Bolt', 'E-Scooter'];
+        }
+        if (placeName.startsWith('lime')) {
+          return ['Lime', 'E-Scooter'];
+        }
+        if (placeName.startsWith('voi')) {
+          return ['VOI', 'E-Scooter'];
+        }
+        if (placeName.startsWith('bird')) {
+          return ['Bird', 'E-Scooter'];
+        }
+      } else {
+        const operatorName = (place.properties['OPERATOR_NAME'] ?? 'n/a').toLowerCase();
+        if (operatorName.startsWith('bird')) {
+          return ['Bird', 'E-Scooter'];
+        }
+        if (operatorName.startsWith('bolt')) {
+          return ['Bolt', 'E-Scooter'];
+        }
+        if (operatorName.startsWith('lime')) {
+          return ['Lime', 'E-Scooter'];
+        }
+        if (operatorName.startsWith('voi')) {
+          return ['VOI', 'E-Scooter'];
+        }
+      }
+
       if (poiCategory === 'bicycle_rental') {
         if (poi.placeName === 'Publibike') {
-          return ['PubliBike', 'Bike']
+          return ['PubliBike', 'Bike'];
         }
 
         if (poi.placeName === 'Publiebike') {
-          return ['PubliBike', 'E-Bike']
+          return ['PubliBike', 'E-Bike'];
         }
 
         if (poi.placeName === 'Nextbike') {
-          return ['Nextbike', 'Bike']
+          return ['Nextbike', 'Bike'];
         }
       }
 
       if (code.startsWith('2em_cars')) {
-        return ['2EM Car Sharing', 'Car']
+        return ['2EM Car Sharing', 'Car'];
       }
 
       if (code.startsWith('2em_cars_e')) {
-        return ['2EM Car Sharing', 'E-Car']
-      }
-
-      if (code.startsWith('bird-')) {
-        return ['Bird', 'E-Scooter']
-      }
-
-      if (code.startsWith('bolt_')) {
-        return ['Bolt', 'E-Scooter']
+        return ['2EM Car Sharing', 'E-Car'];
       }
 
       if (code.startsWith('carvelo2go')) {
-        return ['Carvelo2go', 'E-CargoBike']
+        return ['Carvelo2go', 'E-CargoBike'];
       }
 
       if (code.startsWith('donkey_kreuzlingen')) {
-        return ['Regivelo.ch', 'Bike']
+        return ['Regivelo.ch', 'Bike'];
       }
 
       if (code.startsWith('donkey_')) {
-        return ['Donkey Republic', 'Bike']
+        return ['Donkey Republic', 'Bike'];
       }
 
       if (code.startsWith('edrivecarsharing')) {
-        return ['E-drive Car Sharing', 'E-Car']
-      }
-
-      if (code.startsWith('bird-platform-partner-jmfleetswl-')) {
-        return ['JM Fleets', 'E-Scooter']
+        return ['E-drive Car Sharing', 'E-Car'];
       }
 
       if (code.startsWith('liemobil_liechtenstein_ebike')) {
-        return ['LIEmobil', 'E-Bike']
-      }
-
-      if (code.startsWith('lime_')) {
-        if (code.includes('_escooter')) {
-          return ['Lime', 'E-Scooter']
-        } else {
-          return ['Lime', 'E-Bike']
-        }
+        return ['LIEmobil', 'E-Bike'];
       }
 
       if (code.startsWith('mobility')) {
-        return ['Mobility', 'Car']
+        return ['Mobility', 'Car'];
       }
 
       if (code.startsWith('emobility')) {
-        return ['Mobility', 'E-Car']
+        return ['Mobility', 'E-Car'];
       }
 
       if (code.startsWith('nextbike_ch')) {
-        return ['Nextbike', 'Bike']
+        return ['Nextbike', 'Bike'];
       }
 
       if (code.startsWith('pickebike_')) {
         if (code.includes('_emoped')) {
-          return ['Pick-e-Bike', 'E-Moped']
+          return ['Pick-e-Bike', 'E-Moped'];
         } else {
-          return ['Pick-e-Bike', 'E-Bike']
+          return ['Pick-e-Bike', 'E-Bike'];
         }
       }
 
       if (code.startsWith('share_birrer_ch')) {
-        return ['Share Birrer', 'E-Car']
+        return ['Share Birrer', 'E-Car'];
       }
 
       if (code.startsWith('sponticar')) {
-        return ['Sponti-Car', 'E-Car']
-      }
-
-      if (code.startsWith('tier')) {
-        return ['TIER', 'E-Scooter']
+        return ['Sponti-Car', 'E-Car'];
       }
 
       if (code.startsWith('tier_ebike')) {
-        return ['TIER', 'E-Bike']
+        return ['TIER', 'E-Bike'];
       }
 
-      if (code.startsWith('voiscooters.com')) {
-        return ['VOI', 'E-Scooter']
-      }
-
-      return [null, 'E-Scooter']
+      return [null, 'E-Scooter'];
     })()
-    const provider = providerData[0]
-    const vehicleType = providerData[1]
+    const provider = providerData[0];
+    const vehicleType = providerData[1];
 
     if (provider === null) {
       console.error('CANT DETECT provider');